refactor: renamed login components and updated imports
